Allow default_text in input dialogs
TextArea allows an argument (``text``) to prefill the widget with text. This change simply exposes it for use in ``input_dialog``. This is nice as default text is often used when giving the user hints or when autofilling previously entered values (e.g. from an old session that has been persisted)